backup.sh: rework to run in container
Full rework of backup.sh to run in a container with envsubst applying
env vars to scripts while piping them to balena ssh.

A temporary rsa key will be generated on container start at it will be
added to balena cloud and removed on cleanup.

This same key will also be temporarily installed on the device host OS
in order to allow rsync to communicate via ssh over the tunnel.

Change-type: major
Signed-off-by: Kyle Harding <kyle@balena.io>