v1.0.0: convert to balena app
As part of v1.0.0 we are moving to an orchestrator/node model
where one app will execute duplicity backups on remote devices.

This commit includes the execute component will normally be called
by a scheduler, but can also be called on demand for advanced usage.

The execute component is a generic wrapper to connect to a remote
device via ssh, run a docker image with all named volumes, and execute
a provided command.

The pipeme component is a script that will have variables substituted
before it is piped over balena ssh to the remote device as commands.

The backup and restore components are just helper scripts that will
read tags from the device before executing the backup or restore
action.

Change-type: major
Signed-off-by: Kyle Harding <kyle@balena.io>